# backend/app/api/v1/routers/rides.py
"""Ride endpoints: request, offers, assignment and lifecycle."""

from __future__ import annotations

import asyncio
import logging
import uuid
from datetime import UTC, datetime
from typing import Annotated

from fastapi import APIRouter, Depends, Query, Response, status
from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker

from app.api.deps import (
    CurrentUserDep,
    PassengerPresenceLeaseStoreDep,
    SessionFactoryDep,
    SettingsDep,
    build_expire_offer_and_complete_scheduled_action,
    get_accept_offer,
    get_cancel_ride,
    get_create_offer,
    get_create_ride_request,
    get_dismiss_open_ride,
    get_edit_ride,
    get_get_ride,
    get_list_offers_for_ride,
    get_list_open_rides,
    get_list_recent_destinations,
    get_list_ride_history,
    get_mark_rider_on_the_way,
    get_passenger_active_ride,
    get_pause_ride_for_edit,
    get_pending_rating_ride,
    get_rate_ride,
    get_reject_offer,
    get_ride_rating,
    get_skip_ride_rating,
    get_update_ride_fare,
    get_update_ride_status,
    get_withdraw_offer,
)
from app.api.v1 import events, presence
from app.api.v1.pagination import decode_cursor
from app.api.v1.schemas.offers import OfferCreate, OfferResponse
from app.api.v1.schemas.ratings import RatingCreate, RatingResponse
from app.api.v1.schemas.rides import (
    CreateRideRequestRequest,
    OpenRidePageResponse,
    RecentDestinationResponse,
    RideEdit,
    RideFareUpdate,
    RideHistoryPageResponse,
    RideRequestResponse,
    RideResponse,
    RideStatusUpdate,
)
from app.application.dto import (
    CreateOfferInput,
    CreateRideRequestInput,
    LocationInput,
    RideDetail,
)
from app.application.use_cases.accept_offer import AcceptOffer
from app.application.use_cases.cancel_ride import CancelRide
from app.application.use_cases.create_offer import CreateOffer
from app.application.use_cases.create_ride_request import CreateRideRequest
from app.application.use_cases.dismiss_open_ride import DismissOpenRide
from app.application.use_cases.edit_ride import EditRide
from app.application.use_cases.get_passenger_active_ride import GetPassengerActiveRide
from app.application.use_cases.get_pending_rating_ride import GetPendingRatingRide
from app.application.use_cases.get_ride import GetRide
from app.application.use_cases.get_ride_rating import GetRideRating
from app.application.use_cases.list_offers_for_ride import ListOffersForRide
from app.application.use_cases.list_open_rides import ListOpenRides
from app.application.use_cases.list_recent_destinations import ListRecentDestinations
from app.application.use_cases.list_ride_history import ListRideHistory
from app.application.use_cases.mark_rider_on_the_way import MarkRiderOnTheWay
from app.application.use_cases.pause_ride_for_edit import PauseRideForEdit
from app.application.use_cases.rate_ride import RateRide
from app.application.use_cases.reject_offer import RejectOffer
from app.application.use_cases.skip_ride_rating import SkipRideRating
from app.application.use_cases.update_ride_fare import UpdateRideFare
from app.application.use_cases.update_ride_status import UpdateRideStatus
from app.application.use_cases.withdraw_offer import WithdrawOffer
from app.domain.entities import RideStatus
from app.domain.ride_policy import OFFER_TTL
from app.infrastructure.config import Settings

router = APIRouter(prefix="/rides", tags=["rides"])
logger = logging.getLogger(__name__)

def _to_location_input(point) -> LocationInput:
    return LocationInput(
        latitude=point.latitude,
        longitude=point.longitude,
        name=point.name,
        address=point.address,
        country_code=point.country_code,
    )


# In-flight offer expiry tasks. They are kept so asyncio's GC
# does not reclaim them before the 30 s (create_task without a reference is fragile).
_EXPIRY_TASKS: set[asyncio.Task[None]] = set()


async def shutdown_expiry_tasks() -> None:
    """Cancel and await the pending expiries when shutting down the API."""
    tasks = set(_EXPIRY_TASKS)
    for task in tasks:
        task.cancel()
    if tasks:
        await asyncio.gather(*tasks, return_exceptions=True)
    _EXPIRY_TASKS.clear()


async def _expire_offer_after(
    offer_id: uuid.UUID,
    session_factory: async_sessionmaker[AsyncSession],
    settings: Settings,
) -> None:
    """Expire the offer after 30 s and notify driver and passenger in real time over WS.

    A deferred task launched when the offer is created, with a new session (the
    request's session is already closed). If by then the offer was accepted/rejected/withdrawn/
    improved, the use case does not touch it (race-safe). It is best-effort: it must never
    break the worker (server restarted, session dropped, etc.).
    """
    try:
        await asyncio.sleep(OFFER_TTL.total_seconds())
        async with session_factory() as session:
            offer = await build_expire_offer_and_complete_scheduled_action(
                session,
                settings,
            ).execute(offer_id, datetime.now(UTC))
        if offer is not None:
            await events.publish_offer_expired(offer)
    except Exception:
        # Direct delivery is best-effort, but the failure must be observable.
        logger.exception("No se pudo vencer la oferta %s", offer_id)


@router.post("", response_model=RideRequestResponse, status_code=status.HTTP_201_CREATED)
async def create_ride(
    body: CreateRideRequestRequest,
    current_user: CurrentUserDep,
    use_case: Annotated[CreateRideRequest, Depends(get_create_ride_request)],
) -> RideRequestResponse:
    ride = await use_case.execute(
        current_user,
        CreateRideRequestInput(
            origin=_to_location_input(body.origin),
            destination=_to_location_input(body.destination),
            service_type=body.service_type,
            fare=body.fare,
            payment_method=body.payment_method,
        ),
    )
    # The request shows up for drivers when the passenger opens their
    # WebSocket connection (presence), not when it is created. See ``passenger_ws``.
    return RideRequestResponse.from_entity(ride)


@router.get("/recent-destinations", response_model=list[RecentDestinationResponse])
async def recent_destinations(
    current_user: CurrentUserDep,
    use_case: Annotated[ListRecentDestinations, Depends(get_list_recent_destinations)],
) -> list[RecentDestinationResponse]:
    locations = await use_case.execute(current_user.id)
    return [RecentDestinationResponse.from_location(loc) for loc in locations]


@router.get("/open", response_model=OpenRidePageResponse)
async def open_rides(
    current_user: CurrentUserDep,
    use_case: Annotated[ListOpenRides, Depends(get_list_open_rides)],
    settings: SettingsDep,
    passenger_presence: PassengerPresenceLeaseStoreDep,
    cursor: Annotated[str | None, Query(max_length=512)] = None,
    limit: Annotated[int, Query(ge=1, le=100)] = 50,
) -> OpenRidePageResponse:
    """Open requests for the (online) driver's vehicle type.

    Only those whose passenger is present (live WS connection): abandoned ones
    are not shown.
    """
    page = await use_case.execute(current_user, decode_cursor(cursor), limit)
    if settings.realtime_shared_presence_enabled:
        if passenger_presence is not None:
            page = await presence.present_rides_shared(page, passenger_presence)
    else:
        page = presence.present_rides(page)
    return OpenRidePageResponse.from_page(page)


@router.post("/{ride_id}/dismiss", status_code=status.HTTP_204_NO_CONTENT)
async def dismiss_open_ride(
    ride_id: uuid.UUID,
    current_user: CurrentUserDep,
    use_case: Annotated[DismissOpenRide, Depends(get_dismiss_open_ride)],
) -> Response:
    """Hide the current version of an open request for this driver."""
    await use_case.execute(current_user, ride_id)
    return Response(status_code=status.HTTP_204_NO_CONTENT)


@router.post(
    "/offers/{offer_id}/accept",
    response_model=RideResponse,
)
async def accept_offer(
    offer_id: uuid.UUID,
    current_user: CurrentUserDep,
    use_case: Annotated[AcceptOffer, Depends(get_accept_offer)],
) -> RideResponse:
    """The passenger accepts an offer: it assigns them the ride (final decision).

    The ride's other live offers are rejected in the same transaction.
    """
    result = await use_case.execute(current_user, offer_id)
    await events.publish_offer_accepted(result)
    return RideResponse.from_detail(result.detail)


@router.post("/offers/{offer_id}/reject", status_code=status.HTTP_204_NO_CONTENT)
async def reject_offer(
    offer_id: uuid.UUID,
    current_user: CurrentUserDep,
    use_case: Annotated[RejectOffer, Depends(get_reject_offer)],
) -> Response:
    """The passenger rejects a specific offer; the driver sees it live."""
    offer = await use_case.execute(current_user, offer_id)
    await events.publish_offer_rejected(offer)
    return Response(status_code=status.HTTP_204_NO_CONTENT)


@router.post("/offers/{offer_id}/withdraw", status_code=status.HTTP_204_NO_CONTENT)
async def withdraw_offer(
    offer_id: uuid.UUID,
    current_user: CurrentUserDep,
    use_case: Annotated[WithdrawOffer, Depends(get_withdraw_offer)],
) -> Response:
    """The driver withdraws (or declines to confirm) their offer; the passenger stops seeing it."""
    offer = await use_case.execute(current_user, offer_id)
    await events.publish_offer_withdrawn_by_driver(offer)
    return Response(status_code=status.HTTP_204_NO_CONTENT)


@router.get("/history", response_model=RideHistoryPageResponse)
async def ride_history(
    current_user: CurrentUserDep,
    use_case: Annotated[ListRideHistory, Depends(get_list_ride_history)],
    status: RideStatus | None = None,
    cursor: Annotated[str | None, Query(max_length=512)] = None,
    limit: Annotated[int, Query(ge=1, le=100)] = 20,
) -> RideHistoryPageResponse:
    """Historial de viajes terminales del usuario (pasajero o conductor)."""
    page = await use_case.execute(current_user, status, decode_cursor(cursor), limit)
    return RideHistoryPageResponse.from_page(page)


@router.get("/me/active", response_model=RideResponse | None)
async def passenger_active_ride(
    current_user: CurrentUserDep,
    use_case: Annotated[GetPassengerActiveRide, Depends(get_passenger_active_ride)],
    session_factory: SessionFactoryDep,
    settings: SettingsDep,
    passenger_presence: PassengerPresenceLeaseStoreDep,
) -> RideResponse | None:
    """The passenger's non-terminal request or ride, to resume the flow."""
    detail = await use_case.execute(current_user)
    if (
        detail is not None
        and detail.ride.status is RideStatus.SEARCHING
        and not detail.ride.paused
    ):
        if (
            settings.realtime_shared_presence_enabled
            and passenger_presence is not None
        ):
            await presence.on_shared_passenger_activity(
                detail.ride.id,
                session_factory,
                passenger_presence,
            )
        else:
            await presence.on_passenger_activity(
                detail.ride.id,
                session_factory,
                settings,
            )
    return RideResponse.from_detail(detail) if detail is not None else None


@router.get("/me/pending-rating", response_model=RideResponse | None)
async def pending_rating_ride(
    current_user: CurrentUserDep,
    use_case: Annotated[GetPendingRatingRide, Depends(get_pending_rating_ride)],
) -> RideResponse | None:
    """The latest completed ride the current user has not rated yet."""
    detail = await use_case.execute(current_user)
    return RideResponse.from_detail(detail) if detail is not None else None


@router.get("/{ride_id}", response_model=RideResponse)
async def get_ride(
    ride_id: uuid.UUID,
    current_user: CurrentUserDep,
    use_case: Annotated[GetRide, Depends(get_get_ride)],
) -> RideResponse:
    """Ride detail for polling (passenger or assigned driver)."""
    detail = await use_case.execute(current_user, ride_id)
    return RideResponse.from_detail(detail)


@router.get("/{ride_id}/rating", response_model=RatingResponse | None)
async def get_ride_rating_response(
    ride_id: uuid.UUID,
    current_user: CurrentUserDep,
    use_case: Annotated[GetRideRating, Depends(get_ride_rating)],
) -> RatingResponse | None:
    rating = await use_case.execute(current_user, ride_id)
    return RatingResponse.from_entity(rating) if rating is not None else None


@router.post(
    "/{ride_id}/rating",
    response_model=RatingResponse,
    status_code=status.HTTP_201_CREATED,
)
async def rate_ride(
    ride_id: uuid.UUID,
    body: RatingCreate,
    current_user: CurrentUserDep,
    use_case: Annotated[RateRide, Depends(get_rate_ride)],
) -> RatingResponse:
    """Rate the other participant after the ride is completed."""
    rating = await use_case.execute(current_user, ride_id, body.score, body.comment)
    return RatingResponse.from_entity(rating)


@router.post("/{ride_id}/rating/skip", status_code=status.HTTP_204_NO_CONTENT)
async def skip_ride_rating(
    ride_id: uuid.UUID,
    current_user: CurrentUserDep,
    use_case: Annotated[SkipRideRating, Depends(get_skip_ride_rating)],
) -> Response:
    """Close the pending rating without changing reputation."""
    await use_case.execute(current_user, ride_id)
    return Response(status_code=status.HTTP_204_NO_CONTENT)


@router.get("/{ride_id}/offers", response_model=list[OfferResponse])
async def list_offers(
    ride_id: uuid.UUID,
    current_user: CurrentUserDep,
    use_case: Annotated[ListOffersForRide, Depends(get_list_offers_for_ride)],
) -> list[OfferResponse]:
    """Pending offers the passenger received for their ride."""
    offers = await use_case.execute(current_user, ride_id)
    return [OfferResponse.from_detail(detail) for detail in offers]


@router.post(
    "/{ride_id}/offers",
    response_model=OfferResponse,
    status_code=status.HTTP_201_CREATED,
)
async def create_offer(
    ride_id: uuid.UUID,
    body: OfferCreate,
    current_user: CurrentUserDep,
    use_case: Annotated[CreateOffer, Depends(get_create_offer)],
    session_factory: SessionFactoryDep,
    settings: SettingsDep,
) -> OfferResponse:
    """The driver offers on a ride (accept at the price, counter-offer or improve)."""
    result = await use_case.execute(
        current_user,
        ride_id,
        CreateOfferInput(
            accept_at_fare=body.accept_at_fare,
            price=body.price,
            eta_min=body.eta_min,
            expected_pool_version=body.expected_pool_version,
        ),
    )
    if result.superseded_offer_id is not None:
        # Offer improvement: withdraw the old card and announce the new one.
        await events.publish_offer_superseded(result.superseded_offer_id, result.detail)
    else:
        await events.publish_offer_created(result.detail)
    # In shadow mode the timer is kept to compare the dual write. In live mode, the
    # durable action was already inserted in the same transaction as the offer.
    if settings.scheduled_actions_mode != "live":
        task = asyncio.create_task(
            _expire_offer_after(
                result.detail.offer.id,
                session_factory,
                settings,
            )
        )
        _EXPIRY_TASKS.add(task)
        task.add_done_callback(_EXPIRY_TASKS.discard)
    return OfferResponse.from_detail(result.detail)


@router.post("/{ride_id}/rider-on-the-way", response_model=RideResponse)
async def mark_rider_on_the_way(
    ride_id: uuid.UUID,
    current_user: CurrentUserDep,
    use_case: Annotated[MarkRiderOnTheWay, Depends(get_mark_rider_on_the_way)],
) -> RideResponse:
    """Persist a passenger's pickup notice and deliver it to both participants."""
    detail = await use_case.execute(current_user, ride_id)
    await events.publish_ride_status(detail)
    return RideResponse.from_detail(detail)


@router.patch("/{ride_id}/status", response_model=RideResponse)
async def update_status(
    ride_id: uuid.UUID,
    body: RideStatusUpdate,
    current_user: CurrentUserDep,
    use_case: Annotated[UpdateRideStatus, Depends(get_update_ride_status)],
) -> RideResponse:
    """The assigned driver advances the ride's status."""
    detail = await use_case.execute(current_user, ride_id, body.status)
    await events.publish_ride_status(detail)
    return RideResponse.from_detail(detail)


@router.patch("/{ride_id}/fare", response_model=RideResponse)
async def update_fare(
    ride_id: uuid.UUID,
    body: RideFareUpdate,
    current_user: CurrentUserDep,
    use_case: Annotated[UpdateRideFare, Depends(get_update_ride_fare)],
) -> RideResponse:
    """The passenger adjusts their offer while drivers are being searched."""
    result = await use_case.execute(current_user, ride_id, body.fare)
    await events.publish_ride_republished(result)
    return RideResponse.from_detail(result.detail)


@router.post("/{ride_id}/cancel", response_model=RideResponse)
async def cancel_ride(
    ride_id: uuid.UUID,
    current_user: CurrentUserDep,
    use_case: Annotated[CancelRide, Depends(get_cancel_ride)],
) -> RideResponse:
    """Cancel the ride (passenger or assigned driver) before it starts."""
    result = await use_case.execute(current_user, ride_id)
    await events.publish_ride_cancelled(result)
    return RideResponse.from_detail(result.detail)


@router.post("/{ride_id}/pause-edit", response_model=RideResponse)
async def pause_ride_for_edit(
    ride_id: uuid.UUID,
    current_user: CurrentUserDep,
    use_case: Annotated[PauseRideForEdit, Depends(get_pause_ride_for_edit)],
) -> RideResponse:
    """Pause the request to edit it (Modify): hide it from the pool and withdraw offers."""
    result = await use_case.execute(current_user, ride_id)
    await events.publish_ride_paused(result)
    return RideResponse.from_detail(
        RideDetail(ride=result.ride, rider=current_user, driver=None, accepted_offer=None)
    )


@router.patch("/{ride_id}", response_model=RideResponse)
async def edit_ride(
    ride_id: uuid.UUID,
    body: RideEdit,
    current_user: CurrentUserDep,
    use_case: Annotated[EditRide, Depends(get_edit_ride)],
) -> RideResponse:
    """Save the changes to a paused request and publish it to the pool again."""
    result = await use_case.execute(
        current_user,
        ride_id,
        CreateRideRequestInput(
            origin=_to_location_input(body.origin),
            destination=_to_location_input(body.destination),
            service_type=body.service_type,
            fare=body.fare,
            payment_method=body.payment_method,
        ),
    )
    await events.publish_ride_republished(result)
    return RideResponse.from_detail(result.detail)