# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateRideRequestRequest": {
        "properties": {
          "destination": {
            "$ref": "#/components/schemas/PointInputSchema"
          },
          "fare": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,8}|(?=[\\d.]{1,11}0*$)\\d{0,8}\\.\\d{0,2}0*$)",
                "type": "string"
              }
            ],
            "title": "Fare"
          },
          "origin": {
            "$ref": "#/components/schemas/PointInputSchema"
          },
          "payment_method": {
            "$ref": "#/components/schemas/PaymentMethod",
            "default": "cash"
          },
          "service_type": {
            "$ref": "#/components/schemas/ServiceType"
          }
        },
        "required": [
          "origin",
          "destination",
          "service_type",
          "fare"
        ],
        "title": "CreateRideRequestRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PaymentMethod": {
        "description": "Payment method chosen for the ride.\n\nFor now the app supports QR payment and cash.",
        "enum": [
          "qr",
          "cash"
        ],
        "title": "PaymentMethod",
        "type": "string"
      },
...
      "ServiceType": {
        "description": "Service requested by a passenger (``MOVING`` is a house/office move).",
        "enum": [
          "taxi",
          "moto",
          "delivery",
          "moving"
        ],
        "title": "ServiceType",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "ViajaYa API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/rides": {
      "post": {
        "operationId": "create_ride_api_v1_rides_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRideRequestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RideRequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Ride",
        "tags": [
          "rides"
        ]
      }
    },
    "/api/v1/rides/history": {
      "get": {
        "description": "Historial de viajes terminales del usuario (pasajero o conductor).",
        "operationId": "ride_history_api_v1_rides_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RideStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RideHistoryPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ride History",
        "tags": [
          "rides"
        ]
      }
    },
    "/api/v1/rides/me/active": {
      "get": {
        "description": "The passenger's non-terminal request or ride, to resume the flow.",
        "operationId": "passenger_active_ride_api_v1_rides_me_active_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/RideResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Passenger Active Ride Api V1 Rides Me Active Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Passenger Active Ride",
        "tags": [
          "rides"
        ]
      }
    },
    "/api/v1/rides/me/pending-rating": {
      "get": {
        "description": "The latest completed ride the current user has not rated yet.",
        "operationId": "pending_rating_ride_api_v1_rides_me_pending_rating_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/RideResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Pending Rating Ride Api V1 Rides Me Pending Rating Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pending Rating Ride",
        "tags": [
          "rides"
        ]
      }
    },
    "/api/v1/rides/offers/{offer_id}/accept": {
      "post": {
        "description": "The passenger accepts an offer: it assigns them the ride (final decision).\n\nThe ride's other live offers are rejected in the same transaction.",
        "operationId": "accept_offer_api_v1_rides_offers__offer_id__accept_post",
        "parameters": [
          {
            "in": "path",
            "name": "offer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Offer Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RideResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept Offer",
        "tags": [
          "rides"
        ]
      }
    },
    "/api/v1/rides/offers/{offer_id}/reject": {
      "post": {
        "description": "The passenger rejects a specific offer; the driver sees it live.",
        "operationId": "reject_offer_api_v1_rides_offers__offer_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "offer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Offer Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Offer",
        "tags": [
          "rides"
        ]
      }
    },
    "/api/v1/rides/offers/{offer_id}/withdraw": {
      "post": {
        "description": "The driver withdraws (or declines to confirm) their offer; the passenger stops seeing it.",
        "operationId": "withdraw_offer_api_v1_rides_offers__offer_id__withdraw_post",
        "parameters": [
          {
            "in": "path",
            "name": "offer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Offer Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Withdraw Offer",
        "tags": [
          "rides"
        ]
      }
    },
    "/api/v1/rides/open": {
      "get": {
        "description": "Open requests for the (online) driver's vehicle type.\n\nOnly those whose passenger is present (live WS connection): abandoned ones\nare not shown.",
        "operationId": "open_rides_api_v1_rides_open_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenRidePageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Open Rides",
        "tags": [
          "rides"
        ]
      }
    },
    "/api/v1/rides/recent-destinations": {
      "get": {
        "operationId": "recent_destinations_api_v1_rides_recent_destinations_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RecentDestinationResponse"
                  },
                  "title": "Response Recent Destinations Api V1 Rides Recent Destinations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recent Destinations",
        "tags": [
          "rides"
        ]
      }
    },
...
  }
}